[OSIS-3357] Rename file reporting to excel
	renamed:    base/business/education_groups/reporting.py -> base/business/education_groups/excel.py
	modified:   base/templates/education_group/blocks/dropdown/identification_actions_bar.html
	renamed:    base/tests/business/education_groups/test_reporting.py -> base/tests/business/education_groups/test_excel.py
	renamed:    base/tests/views/education_groups/test_reporting.py -> base/tests/views/education_groups/test_excel.py
	renamed:    base/views/education_groups/reporting.py -> base/views/education_groups/excel.py
	modified:   base/views/education_groups/urls.py
