+++ base/templates/education_group/blocks/dropdown/identification_actions_bar.html
@@ -114,8 +114,8 @@
                 </li>
                 <li class="divider"></li>
                 <li>
-                    <a href="{% url "prerequisites_reporting" education_group_year_pk=education_group_year.pk %}">
-                        <span>{% trans 'Excel cours ayant prérequis' %}</span>
+                    <a href="{% url "education_group_learning_units_prerequisites" education_group_year_pk=education_group_year.pk %}">
+                        <span>{% trans 'Excel prérequis UEs' %}</span>
                     </a>
                 </li>
