+++ base/views/education_groups/urls.py
@@ -26,7 +26,7 @@
 from django.conf.urls import url, include
 
 from base.views import education_group
-from base.views.education_groups import reporting
+from base.views.education_groups import excel as education_group_excel
 from base.views.education_groups.group_element_year.read import pdf_content
 from base.views.education_groups.learning_unit import detail as learning_unit_detail, update as learning_unit_update
 from base.views.education_groups.publication_contact import CreateEducationGroupPublicationContactView, \
@@ -175,10 +175,9 @@ urlpatterns = [
             learning_unit_update.LearningUnitPrerequisite.as_view(),
             name='learning_unit_prerequisite_update'),
     ])),
-    # TODO Rename name of url
     url(
         r'reporting/(?P<education_group_year_pk>[0-9]+)/prerequisites/$',
-        reporting.get_learning_unit_prerequisites_excel,
-        name="prerequisites_reporting"
+        education_group_excel.get_learning_unit_prerequisites_excel,
+        name="education_group_learning_units_prerequisites"
     )
 ]
